--- docker/importer/testdata/expected_patch_basic.diff
@@ -1,35 +1,0 @@
-diff --git a/file/OSV-2017-134.yaml b/file/OSV-2017-134.yaml
-new file mode 100644
-index 0000000..2a0cc23
---- /dev/null
-+++ b/file/OSV-2017-134.yaml
-@@ -0,0 +1,29 @@
-+id: OSV-2017-134
-+package:
-+  name: file
-+  ecosystem: OSS-Fuzz
-+summary: Heap-buffer-overflow in cdf_file_property_info
-+details: |
-+  OSS-Fuzz report: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=1064
-+
-+  Crash type: Heap-buffer-overflow READ 1
-+  Crash state:
-+  cdf_file_property_info
-+  cdf_file_summary_info
-+  cdf_check_summary_info
-+affects:
-+  ranges:
-+  - type: GIT
-+    repo: https://github.com/file/file.git
-+    introduced: 17ee4cf670c363de8d2ea4a4897d7a699837873f
-+    fixed: 19ccebafb7663c422c714e0c67fa4775abf91c43
-+  versions:
-+  - FILE5_29
-+  - FILE5_30
-+modified: '2021-01-01T00:00:00Z'
-+published: '2021-01-15T00:00:24.559102Z'
-+ecosystem_specific:
-+  severity: MEDIUM
-+references:
-+- type: REPORT
-+  url: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=1064
